Make migrations idempotent

+++ db/migrate/20220315172902_community_add_score_to_directory_items.rb
@@ -1,10 +1,12 @@
 # frozen_string_literal: true
 class CommunityAddScoreToDirectoryItems < ActiveRecord::Migration[6.1]
   def up
-    add_column :directory_items, :gamification_score, :integer, default: 0
+    unless column_exists?(:directory_items, :gamification_score)
+      add_column :directory_items, :gamification_score, :integer, default: 0
+    end
   end
 
   def down
-    remove_column :directory_items, :gamification_score
+    remove_column :directory_items, :gamification_score if column_exists?(:directory_items, :gamification_score)
   end
 end
